requirement.txt
Flask
opencv-python
numpy
qrcode
pandas
reportlab
Pillow
